requirements.txt
langchain>=0.2
langchain-community
langchain_openai
nicegui
tortoise-orm
